Refactor to use new metadata structure and always with tagdiff file
Rework logic to always utilise a tagdiff file as input (even for master branch).
Check that every tag to be imported is in git.
Check that every tag to be imported has not yet been processed for branch.
Commit is made after every tag is processed; final tag for release number.

Make recursive delete safe if directory is missing.