Move changelog diff to atutils; add "ignore_fail" option to check_output_with_retry
Make changelog_diff more useful (feed with package path) and ensure that it will
work with staged changes (i.e., result of "git checkout TAG PATH").
Do not produce a diff if the ChangeLog is new (i.e., first import).

"ignore_fail" option added to check_output_with_retry to cope with the fact
that an orphaned branch with has no refs on it (i.e., HEAD does not exist).