Prevent master branch from importing releases that would go back in time
If the onlyforward option is set, then we need to preprocess the list of tag content
files in order to ensure that we never jump back in time to a previous release series
when making the master branch. This is because the earlier release series will be
branched off from and the later releases on that series really only make sense
for the branch (thus time travel needs to be checked going backwards). Function
backskip_filter() implements this.

Move branch_exists() function to atutils.