Fix svnpull.py --files option
The --files option to svnpull.py was buggy, because the clobbered
update of the whole package tree means that files which should have
been left alone were mistakenly deleted. Now, if the --files option
is used then the target directory in git is left alone and only
explicitly imported files from SVN are copied in.

When the normal package copy mode is used then the updated
doxygen file (packagedoc.h) is saved from being deleted if
there is no packagedoc.h in the SVN pulled area. This is because
of the doxygen update to replace mainpage.h with packagedoc.h
(https://gitlab.cern.ch/atlas/athena/merge_requests/171).